fix issue in the fuel and teu dwt calcuation

+++ src/main/java/it/spaghettisource/navaltrader/ui/model/BuyShipTableRow.java
@@ -13,11 +13,11 @@ public class BuyShipTableRow {
 	private int hull;	
 	private int maxTeu;		
 	private int maxDwt;		
-	private int maxFuel;		
+	private double maxFuel;		
 	private int maxSpeed;	
 	private double operatingCost;
 	
-	public BuyShipTableRow(String shipClass, String type, double price, int hull, int maxDwt,int maxTeu, double operatingCost, int maxFuel, int maxSpeed) {
+	public BuyShipTableRow(String shipClass, String type, double price, int hull, int maxDwt,int maxTeu, double operatingCost, double maxFuel, int maxSpeed) {
 		super();
 		this.shipClass = shipClass;
 		this.model = type;
@@ -58,7 +58,7 @@ public class BuyShipTableRow {
 		return maxTeu;
 	}
 
-	public int getMaxFuel() {
+	public double getMaxFuel() {
 		return maxFuel;
 	}
